Add files via upload
This project features an extensive Exploratory Data Analysis (EDA) of the Zomato restaurant dataset, starting with data integration by merging the main data with a country code file to enable geographical insights. The analysis identifies key global contributors like India, the US, and the UK, and maps countries to their respective currencies. A major finding reveals that online delivery service is exclusively offered in India and the UAE. Furthermore, the project delves into rating distribution, associating aggregate scores with rating text and color, and highlights the significant issue of "Not rated" (0.0 score) restaurants, which are most prevalent in India. Finally, the analysis explores city-level distribution to understand Zomato's operational concentration.